pyproject.toml
[project]
name = "callintel"
version = "0.1.0"
description = "Local AI pipeline for analysing recorded phone calls"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.0",
    "torch>=2.2.0",
    "transformers>=4.45.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "python-dotenv>=1.0.1",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.20.0",
    "scikit-learn>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4", "mypy>=1.8"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
